refactor: remove SelectCategory component and its related items
- Deleted SelectCategory, SelectCategoryItem, SelectCategoryBadge, and SelectCategoryTrigger components from the products categories module to enhance code maintainability and reduce complexity in the product selection UI.
